register/signin form, onPressTicket
