+++ lib/models/ticketPageModel.dart
@@ -1,16 +1,11 @@
 import 'package:flutterflow_ui/flutterflow_ui.dart';
 import 'package:flutter/material.dart';
-
-import '../widgets/ticketPageWidget.dart';
+import '../screens/home/ticketPageScreen.dart';
 
 class TicketPageModel extends FlutterFlowModel<TicketPageWidget> {
-  ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
-  // State field(s) for RatingBar widget.
   double? ratingBarValue;
-
-  /// Initialization and disposal methods.
 
   @override
   void initState(BuildContext context) {}
@@ -19,8 +14,4 @@ class TicketPageModel extends FlutterFlowModel<TicketPageWidget> {
   void dispose() {
     unfocusNode.dispose();
   }
-
-  /// Action blocks are added here.
-
-  /// Additional helper methods are added here.
 }
